requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docx==0.2.4
drf-yasg==1.21.5
et-xmlfile==1.1.0
filelock==3.10.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.24.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipenv==2023.3.20
platformdirs==3.2.0
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
